move about auth defs to app directories auth.py
