+++ flaskr/app/auth.py
@@ -1,0 +1,59 @@
+from flask import Flask, flash, redirect, render_template, request, session, url_for
+from werkzeug.security import check_password_hash, generate_password_hash
+
+
+@app.route('/register', methods=('GET', 'POST'))
+def register():
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+
+        if not username:
+            error = 'Username is required.'
+        elif not password:
+            error = 'Password is required.'
+
+        user = User.query.filter_by(name=username).first()
+
+        if user is None:
+            new_user = User(name=username, password=generate_password_hash(password))
+            db.session.add(new_user)
+            db.session.commit()
+
+            return redirect(url_for('login'))
+
+        error = 'User {} is already registered.'.format(username)
+
+        flash(error)
+
+    return render_template('auth/register.html')
+
+
+@app.route('/login', methods=('GET', 'POST'))
+def login():
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        error = None
+
+        user = User.query.filter_by(name=username).first()
+
+        if user is None:
+            error = 'Incorrect username.'
+        elif not check_password_hash(user.password, password):
+            error = 'Incorrect password.'
+
+        if error is None:
+            session.clear()
+            session['user_id'] = user.id
+            return redirect(url_for('index'))
+
+        flash(error)
+
+    return render_template('auth/login.html')
+
+
+@app.route('/logout')
+def logout():
+    session.clear()
+    return redirect(url_for('login'))
